Struct chumsky_branch::Branch

pub struct Branch<'a, T, E> { /* private fields */ }

The return type of branch.

Implementations

impl<'a, T, E> Branch<'a, T, E>where
E: Error<char> + 'a,

pub fn or_branch<P>(self, begin: &'a str, parser: P) -> Selfwhere
P: Parser<char, T, Error = E> + 'a,

Branch into another parser if the first branch didn’t match.

pub fn or_else<B>(self, branch: B) -> Box<dyn Parser<char, T, Error = E> + 'a>where
B: Fn(String) -> T + 'a,
T: 'a,

Add an else branch if non of the branches succeeded.